fix update and create avartar

+++ controller/actor.controller.js
@@ -1,13 +1,10 @@
 const db = require("../models/init-models");
 const { Op } = require("sequelize");
-const {
-  responseInValid,
-  reponseSuccess,
-  responseSuccessWithData,
-} = require("../helper/ResponseRequests");
+const { responseInValid, reponseSuccess, responseSuccessWithData } = require("../helper/ResponseRequests");
 const path = require("path");
 const root = path.resolve("./");
 const { removeDir } = require("../helper/file");
+const fs = require("fs");
 const get = async (req, res) => {
   let filter = {};
   const { search } = req.query;
@@ -35,16 +32,33 @@ const getById = async (req, res) => {
 };
 
 const create = async (req, res) => {
-  await db.actors.create(req.body);
+  const actor = await db.actors.create(req.body);
+  if (req.body["avartar"]) {
+    const addTail = req.destination.concat("/", `actor-${actor.actor_id}-${req.originalname}`);
+    fs.rename(req.body["avartar"], addTail, (err) => {
+      if (err) console.log(err);
+    });
+    req.body["avartar"] = addTail;
+    actor.update(req.body);
+  }
   return reponseSuccess({ res });
 };
 
 const update = async (req, res) => {
   const actor = await db.actors.findOne({ where: { actor_id: req.params.id } });
   if (!actor) return responseInValid({ res, message: "not found actor" });
-  if (req.body["avartar"] && actor.avartar && actor.avartar !== req.body["avartar"]) {
+  // if (req.body["avartar"] && actor.avartar && actor.avartar !== req.body["avartar"]) {
+  //   const pathOldImage = `${root}/${actor.avartar}`;
+  //   removeDir({ dir: pathOldImage });
+  // }
+  if (req.body["avartar"]) {
+    const addTail = req.destination.concat("/", `actor-${actor.actor_id}-${req.originalname}`);
     const pathOldImage = `${root}/${actor.avartar}`;
-    removeDir({ dir: pathOldImage });
+    fs.rename(req.body["avartar"], addTail, (err) => {
+      if (err) console.log(err);
+    });
+    req.body["avartar"] = addTail;
+    if (actor.avartar) removeDir({ dir: pathOldImage });
   }
   await actor.update(req.body);
   return responseSuccessWithData({ res, data: actor });
